requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
mysqlclient==2.2.0
Pillow==10.1.0
razorpay==1.4.0
requests==2.31.0
python-decouple==3.8
django-phonenumber-field==7.1.0
phonenumbers==8.13.25
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0

